requirements.txt
amqp==5.0.6
aniso8601==9.0.1
anyjson==0.3.3
appdirs==1.4.4
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.52
botocore==1.20.52
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
dj-database-url==0.5.0
Django==3.2
django-admin-honeypot==1.1.0
django-appconf==1.0.4
django-compressor==2.4
django-cors-headers==3.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-etc==1.3.0
django-extensions==3.1.2
django-hitcount==1.3.3
django-hosts==4.0
django-js-asset==1.2.2
django-meta==2.0.0
django-mptt==0.12.0
django-postgrespool2==2.0.1
django-ranged-response==0.2.0
django-redis==4.12.1
django-robots==4.0
django-rosetta==0.9.6
django-simple-captcha==0.5.14
django-storages==1.11.1
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
future==0.18.2
greenlet==1.0.0
gTTS==2.2.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
kombu==5.0.2
lxml==4.6.3
lz4==3.1.3
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
polib==1.1.1
promise==2.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
rcssmin==1.0.6
redis==3.5.3
regex==2021.4.4
reportlab==3.5.67
requests==2.25.1
rjsmin==1.1.0
Rx==3.1.1
s3transfer==0.3.7
singledispatch==3.6.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.8
sqlparse==0.4.1
stripe==2.56.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
